Разделит код из main.js на модули
