+++ js/data.js
@@ -1,0 +1,27 @@
+const NUMBER_PHOTO_DESCRIPTION_CREATED = 25;
+const COMMENT_COUNT = 30;
+const AVATAR_COUNT = 6;
+
+const NAMES = [
+  'Алена',
+  'Иван',
+  'Мария',
+  'Виктор',
+  'Юлия',
+  'Альбина',
+  'Валерий',
+  'Георгий ',
+  'Денис',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.Как можно было поймать такой неудачный момент?!'
+];
+
+export { NUMBER_PHOTO_DESCRIPTION_CREATED, COMMENT_COUNT, AVATAR_COUNT, NAMES, MESSAGES };
+
